+++ src/main/java/com/github/dsipaint/AMGN/entities/plugins/intrinsic/operators/OpRemoveListener.java
@@ -2,6 +2,7 @@ package com.github.dsipaint.AMGN.entities.plugins.intrinsic.operators;
 
 import java.io.IOException;
 
+import com.github.dsipaint.AMGN.entities.Guild;
 import com.github.dsipaint.AMGN.entities.GuildNetwork;
 import com.github.dsipaint.AMGN.io.IOHandler;
 
@@ -19,10 +20,10 @@ public final class OpRemoveListener extends ListenerAdapter
 		String msg = e.getMessage().getContentRaw();
 		String[] args = msg.split(" ");
 		
-		if(args[0].equalsIgnoreCase(GuildNetwork.getPrefix(e.getGuild().getIdLong()) + "removeoperator") && GuildNetwork.operators.contains(e.getMember().getIdLong()))
+		if(args[0].equalsIgnoreCase(GuildNetwork.getPrefix(e.getGuild().getIdLong()) + "removeoperator") && GuildNetwork.isOperator(e.getAuthor()))
 		{
-			//valid user id and is a real user
-			if(args[1].matches("\\d{18}") && e.getGuild().getMemberById(args[1]) != null)
+			//valid user id and is a real user/role
+			if(args[1].matches(GuildNetwork.ID_REGEX) && (e.getGuild().getMemberById(args[1]) != null || e.getGuild().getRoleById(args[1]) != null))
 			{
 				//if user is not already opped
 				if(GuildNetwork.operators.contains(Long.parseLong(args[1])))
@@ -37,15 +38,16 @@ public final class OpRemoveListener extends ListenerAdapter
 						e1.printStackTrace();
 					}
 					GuildNetwork.sendToModlogs(e.getGuild().getIdLong(), e.getAuthor().getAsTag() + " removed "
-							+ e.getGuild().getMemberById(args[1]).getUser().getAsTag() + " as an operator.");
+							+ (e.getGuild().getMemberById(args[1]) != null ? e.getGuild().getMemberById(args[1]).getUser().getAsTag() : (e.getGuild().getRoleById(args[1]).getName() + "role")) 
+							+ " as an operator.");
 					
 					return;
 				}
 				else
-					e.getChannel().sendMessage("User is already not an operator!").queue();
+					e.getChannel().sendMessage("User or role is already not an operator!").queue();
 			}
 			else
-				e.getChannel().sendMessage("Invalid user id").queue();
+				e.getChannel().sendMessage("Invalid id").queue();
 			
 			return;
 		}